+++ src/context.js
@@ -5,10 +5,12 @@ export const margin = 0;
 export const newline = 1;
 export const indent = 2;
 export const dedent = 3;
+export const space = 4;
 
 export class Context {
 	#visitors;
 	#commands;
+	#has_newline = false;
 
 	multiline = false;
 
@@ -35,8 +37,12 @@ export class Context {
 	}
 
 	newline() {
-		this.multiline = true;
+		this.#has_newline = true;
 		this.#commands.push(newline);
+	}
+
+	space() {
+		this.#commands.push(space);
 	}
 
 	/**
@@ -44,6 +50,10 @@ export class Context {
 	 */
 	append(context) {
 		this.#commands.push(context.#commands);
+
+		if (this.#has_newline) {
+			this.multiline = true;
+		}
 	}
 
 	/**
@@ -58,6 +68,10 @@ export class Context {
 			this.location(node.loc.end.line, node.loc.end.column);
 		} else {
 			this.#commands.push(content);
+		}
+
+		if (this.#has_newline) {
+			this.multiline = true;
 		}
 	}