+++ test/samples/comment-after-function-argument/input.js
@@ -1,0 +1,3 @@
+foo(function bar() {
+	// logic goes here
+}, true); // trailing comment