+++ test/samples/comment-between-blocks/expected.js
@@ -1,0 +1,7 @@
+// if true, yep
+if (condition) {
+	console.log('yep');
+} else // otherwise nope
+{
+	console.log('nope');
+}