+++ test/sandbox/index.js
@@ -1,9 +1,9 @@
 // @ts-check
 import { print } from '../../src/index.js';
 import { load } from '../common.js';
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
+import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 import ts from '../../src/languages/ts/index.js';
 
 const dir = path.resolve(fileURLToPath(import.meta.url), '..');